Refactor list column hook to use pluginOptions
