+++ admin/src/components/index.js
@@ -1,4 +1,4 @@
 export { default as CopyLinkButton } from './CopyLinkButton';
 export { default as EditViewRightLinks } from './EditViewRightLinks';
-export { default as ListViewTableCell } from './ListViewTableCell';
+export { default as ListViewColumn } from './ListViewColumn';
 export { default as PreviewButton } from './PreviewButton';
