+++ server/config.js
@@ -8,7 +8,6 @@ const { PREVIEW_WINDOW_NAME } = require('./constants');
 module.exports = {
   default: {
     contentTypes: [],
-    injectListViewColumn: true,
     openTarget: PREVIEW_WINDOW_NAME,
   },
   validator: (config) => {
